task: add duckdb submodules
The purpose of this repository is to coalesce duckdb extensions for
developing a cohesive duckdb engine for skytether. The first step is to
add appropriate submodules (under "subprojects" so that I can make them
play nice in a meson project).

Each submodule is set to use a branch that knows it will be part of this
repository, so they should have their own subprojects removed (just to
make it cleaner and less confusing).