task: updating duckdb submodules
The main changes are in the substrait extension to use mohair-substrait
as a dependency rather than vendoring generated substrait code. Then,
updates to duckdb and duckdb-arrow are to accommodate changes upstream
related to shared_ptr creation (`make_shared_ptr`) and clarifying its
usage (`std::shared_ptr` instead of `duckdb::shared_ptr`).

Also, due to the addition of dependencies to the substrait extension,
some changes to duckdb was necessary (where a cmake function is defined)