task: adding initial impl of skytether storage ext
This defines a storage extension for skytether that will implement an
interface to skytether storage via a custom catalog. This is to
accommodate substrait plans that name a skytether partition (or slice)
as a base table.

This is an alternate path to using an ArrowScan TableFunction to scan an
arrow file and return the data as a result set.